+++ source/renderer/app/containers/loading/SyncingConnectingPage.js
@@ -36,6 +36,7 @@ export default class LoadingSyncingConnectingPage extends Component<Props> {
       isNewAppVersionLoaded,
     } = stores.nodeUpdate;
     const { hasLoadedCurrentLocale, hasLoadedCurrentTheme } = stores.profile;
+    const { isLoadingWallets } = stores.wallets;
     const { toggleNewsFeed } = this.props.actions.app;
     const { unread } = stores.newsFeed.newsFeedData;
     const hasUnreadNews = unread.length > 0;
@@ -49,6 +50,7 @@ export default class LoadingSyncingConnectingPage extends Component<Props> {
         isSynced={isSynced}
         isConnecting={!isConnected}
         isSyncing={isConnected && !isSynced}
+        isLoadingWallets={isLoadingWallets}
         isNodeStopping={isNodeStopping}
         isNodeStopped={isNodeStopped}
         isNotEnoughDiskSpace={isNotEnoughDiskSpace}