+++ source/renderer/app/stores/WalletsStore.js
@@ -738,6 +738,12 @@ export default class WalletsStore extends Store {
     return !!this.active;
   }
 
+  @computed get isLoadingWallets(): boolean {
+    return (
+      this.walletsRequest.result === null && this.walletsRequest.isExecuting
+    );
+  }
+
   @computed get hasLoadedWallets(): boolean {
     return this.walletsRequest.wasExecuted;
   }
@@ -877,8 +883,8 @@ export default class WalletsStore extends Store {
   };
 
   _pollRefresh = async () => {
-    const { isSynced } = this.stores.networkStatus;
-    return isSynced && this.refreshWalletsData();
+    const { isConnected } = this.stores.networkStatus;
+    return isConnected && this.refreshWalletsData();
   };
 
   _updateActiveWalletOnRouteChanges = () => {