+++ storybook/stories/nodes/syncing/SyncingConnecting.stories.js
@@ -27,6 +27,7 @@ export const DefaultSyncingConnectingStory = (props: {
     isSynced={boolean('isSynced', false)}
     isConnecting={boolean('isConnecting', true)}
     isSyncing={boolean('isSyncing', false)}
+    isLoadingWallets={boolean('isLoadingWallets', false)}
     isNodeStopping={boolean('isNodeStopping', false)}
     isNodeStopped={boolean('isNodeStopped', false)}
     isTlsCertInvalid={boolean('isTlsCertInvalid', false)}
@@ -68,6 +69,7 @@ export const ConnectivityIssuesSyncingConnectingStory = (props: {
     isSynced={false}
     isConnecting
     isSyncing={false}
+    isLoadingWallets={false}
     isNodeStopping={false}
     isNodeStopped={false}
     isTlsCertInvalid={false}
@@ -109,6 +111,7 @@ export const SyncIssuesSyncingConnectingStory = (props: {
     isSynced={false}
     isConnecting={false}
     isSyncing
+    isLoadingWallets={false}
     isNodeStopping={false}
     isNodeStopped={false}
     isTlsCertInvalid={false}